% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_practical.R
\name{adm.apply.seed}
\alias{adm.apply.seed}
\title{Assumes the existance of farmdf in an enclosing environment}
\usage{
adm.apply.seed(s, df, newgen = 0)
}
\arguments{
\item{s}{index of return vector}

\item{df}{data frame with farm coordinates}

\item{newgen}{indicator variable for whether there is a new generation (0=no, 1=yes)}
}
\value{
Essentially a void function so returns a 0 for successful execution
}
\description{
Assumes the existance of farmdf in an enclosing environment
}
