% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_practical.R
\name{adm.load.sheep.prep.sim}
\alias{adm.load.sheep.prep.sim}
\title{sets up data set for simulation}
\usage{
adm.load.sheep.prep.sim(df = farmdat)
}
\arguments{
\item{df}{data frame with input data for simulation}
}
\value{
data frame with additional columns 'seqid' and 'g_i'
}
\description{
sets up data set for simulation
}
