% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_practical.R
\name{adm.plot.farm.and.seeds}
\alias{adm.plot.farm.and.seeds}
\title{Function to see farms randomly near to the coast}
\usage{
adm.plot.farm.and.seeds(tab, vecseeds)
}
\arguments{
\item{tab}{data frame with farm coordinates}

\item{vecseeds}{vector of seeds}

\item{nseed}{number of seeds}

\item{sdsq}{vector of bounds for easting and northing coordinates}

\item{tab}{data frame with farm coordinates}
}
\value{
the value of seeds as UIDs

plot of randomly selected farms near the coast
}
\description{
Function to see farms randomly near to the coast
}
