% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_practical.R
\name{adm.plot.gen}
\alias{adm.plot.gen}
\title{Function to plot the generations of the simulation}
\usage{
adm.plot.gen(datf, gens = c(0, 1, 2, 3), cols = c("red", "blue",
  "green", "orange"))
}
\arguments{
\item{datf}{data frame with farm coordinates and generation times}

\item{gens}{vector of generations to be plotted}

\item{cols}{vector of point colors for each generation (should be same length as gens vector)}
}
\value{
plot of the modelled disease spread with a color indicating a different generation
}
\description{
Function to plot the generations of the simulation
}
