% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_practical.R
\name{adm.seed.sheep}
\alias{adm.seed.sheep}
\title{Function to see farms randomly near to the coast}
\usage{
adm.seed.sheep(tab, nseed = 1, sdsq = c(0, 1e+07, 0, 1e+06))
}
\arguments{
\item{tab}{data frame with farm coordinates}

\item{nseed}{number of seeds}

\item{sdsq}{vector of bounds for easting and northing coordinates}
}
\value{
the value of seeds as UIDs
}
\description{
Function to see farms randomly near to the coast
}
