% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_practical.R
\name{adm.setup.aux.mat}
\alias{adm.setup.aux.mat}
\title{creates distance matrix}
\usage{
adm.setup.aux.mat(df)
}
\arguments{
\item{df}{data frame}
}
\value{
distance matrix
}
\description{
creates distance matrix
}
