% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_practical.R
\name{adm.sheep.params}
\alias{adm.sheep.params}
\title{Function to setup a parameter object for the smv simulations}
\usage{
adm.sheep.params()
}
\value{
vector of parameter values
}
\description{
Function to setup a parameter object for the smv simulations
}
