% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\name{augment_data}
\alias{augment_data}
\title{function to augment data}
\usage{
augment_data(data)
}
\arguments{
\item{data}{data frame with a single row, containing the columns 
"time.onset.donor": the observed time of symptom onset for the donor;
"recipient.infected": 1 if the recipient was infected, 0 if not;
"time.infection.recipient": the inferred time at which the recipient was infected;
"time.onset.recipient": the observed time of symptom onset for the recipient}
}
\value{
the same data frame but with a newly generated value for time.infection.recipient
}
\description{
function to augment data
}
