% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\name{calculateLogLik_incubation}
\alias{calculateLogLik_incubation}
\title{calculate log likelihood for the model of incubation only, summed across clusters}
\usage{
calculateLogLik_incubation(p, data)
}
\arguments{
\item{p}{named vector of parameter values.  Contains element "m.incub", the mean
incubation period}

\item{data}{data frame where each row represents a donor-recipient pair.
It contains the column "time.onset.donor" which is the time of symptom onset 
for the donor, i.e. the incubation period for the donor.}
}
\value{
the log likelihood
}
\description{
calculate log likelihood for the model of incubation only, summed across clusters
}
