% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\name{calculateLogLikCluster_incubation}
\alias{calculateLogLikCluster_incubation}
\title{calculate the contribution of a cluster to the log likelihood, for the model
of incubation only}
\usage{
calculateLogLikCluster_incubation(p, data)
}
\arguments{
\item{p}{named vector of parameter values.  Contains element "m.incub", the mean
incubation period}

\item{data}{data frame with a single row, containing the column "time.onset.donor"
which is the time of symptom onset for the donor, i.e. the incubation period
for the donor.}
}
\value{
the contribution of the cluster to the log likelihood
}
\description{
calculate the contribution of a cluster to the log likelihood, for the model
of incubation only
}
