% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\name{calculateLogLikCluster_transmission}
\alias{calculateLogLikCluster_transmission}
\title{calculate the contribution of a cluster to the log likelihood, for the model
of incubation and transmission}
\usage{
calculateLogLikCluster_transmission(p, data)
}
\arguments{
\item{p}{named vector of parameter values.  Contains elements "m.incub", the mean
incubation period;"beta", the transmission parameter; and "m.inf", the mean 
infectious period}

\item{data}{data frame with a single row, containing the columns 
"time.onset.donor": the observed time of symptom onset for the donor;
"recipient.infected": 1 if the recipient was infected, 0 if not;
"time.infection.recipient": the inferred time at which the recipient was infected;
"time.onset.recipient": the observed time of symptom onset for the recipient}
}
\value{
the contribution of the cluster to the log likelihood
}
\description{
calculate the contribution of a cluster to the log likelihood, for the model
of incubation and transmission
}
