% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\name{DA_condition}
\alias{DA_condition}
\title{function to decide whether data should be augmented for a cluster}
\usage{
DA_condition(data)
}
\arguments{
\item{data}{data frame with a single row, containing the columns 
"time.onset.donor": the observed time of symptom onset for the donor;
"recipient.infected": 1 if the recipient was infected, 0 if not;
"time.infection.recipient": the inferred time at which the recipient was infected;
"time.onset.recipient": the observed time of symptom onset for the recipient}
}
\value{
TRUE if the data should be augmented, FALSE otherwise
}
\description{
function to decide whether data should be augmented for a cluster
}
