% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\name{f.incub}
\alias{f.incub}
\title{calculate the probability density for a given incubation period}
\usage{
f.incub(d, p)
}
\arguments{
\item{d}{value of incubation at which to calculate probability density.}

\item{p}{named vector of parameter values.  Contains element "m.incub", the mean
incubation period}
}
\value{
the probability density of the incubation period at d given model
parameters
}
\description{
calculate the probability density for a given incubation period
}
