% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\name{H}
\alias{H}
\title{calculate the cumulative hazard of infection at time t}
\usage{
H(t, p)
}
\arguments{
\item{t}{time at which to calculate the hazard of infection.}

\item{p}{named vector of parameter values.  Contains elements "beta", the
transmission parameter; and "m.inf", the mean infectious period}
}
\value{
the cumulative hazard of infection at time t given model parameters
}
\description{
calculate the cumulative hazard of infection at time t
}
