% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\name{logRatioProposal_DA}
\alias{logRatioProposal_DA}
\title{function to return the log ratio of the proposal distribution at the new and
old values, for the data augmentation}
\usage{
logRatioProposal_DA(newValues, oldValues)
}
\arguments{
\item{newValues}{proposed values for augmented data}

\item{oldValues}{old values for augmented data}
}
\value{
the log ratio of the proposal distribution at the new and
old values
}
\description{
function to return the log ratio of the proposal distribution at the new and
old values, for the data augmentation
}
