% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\name{run_MCMC}
\alias{run_MCMC}
\title{run MCMC algorithm to fit parameters to data}
\usage{
run_MCMC(fit_parnames, nbIteration, randomWalkSD, currentParameter, data)
}
\arguments{
\item{fit_parnames}{vector of names of parameter(s) to be fitted}

\item{nbIteration}{number of iterations for which to run MCMC}

\item{randomWalkSD}{vector containing standard deviations of proposal
distributions for each fitted parameter}

\item{currentParameter}{named vector of parameter values at which to start fitting}

\item{data}{data for fitting, where each row is a cluster}
}
\value{
list with the elements
storedParameter: matrix of fitted parameter values for each iteration
logLik: log likelihood for each iteration
acceptance_rate: proportion of proposed values that were accepted, for each
fitted parameter
}
\description{
run MCMC algorithm to fit parameters to data
}
