% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_practical.R
\name{adm.plot.farm.and.seeds}
\alias{adm.plot.farm.and.seeds}
\title{Function to plot farms randomly near to the coast}
\usage{
adm.plot.farm.and.seeds(tab, vecseeds)
}
\arguments{
\item{tab}{data frame with farm coordinates}

\item{vecseeds}{vector of seeds}
}
\value{
plot of randomly selected farms near the coast
}
\description{
Function to plot farms randomly near to the coast
}
