% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\name{h}
\alias{h}
\title{calculate the hazard of infection at time t}
\usage{
h(t, p)
}
\arguments{
\item{t}{time at which to calculate the hazard of infection.}

\item{p}{named vector of parameter values.  Contains elements "beta", the
transmission parameter; and "m.inf", the mean infectious period}
}
\value{
the hazard of infection at time t given model parameters
}
\description{
calculate the hazard of infection at time t
}
